+++ CHANGES.rst
@@ -15,6 +15,8 @@ Migrations
 Changes
 ~~~~~~~
 
+- Add workaround for andymccurdy/redis-py#633.
+
 - Unify v1 and v2 parse error responses to v2 format.
 
 - Batch key queries depending on a per-model batch size.